NUTCH-1274 Fix [cast] javac warnings
git-svn-id: https://svn.apache.org/repos/asf/nutch/trunk@1432469 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/nutch/crawl/CrawlDatum.java
@@ -457,7 +457,7 @@ public class CrawlDatum implements WritableComparable<CrawlDatum>, Cloneable {
     int res = 0;
     if (signature != null) {
       for (int i = 0; i < signature.length / 4; i += 4) {
-        res ^= (int)(signature[i] << 24 + signature[i+1] << 16 +
+        res ^= (signature[i] << 24 + signature[i+1] << 16 +
                 signature[i+2] << 8 + signature[i+3]);
       }
     }
